Feat/udf (#22)
* output search description by llm

* output no more than 3 word
